requirements.txt
ai21==1.0.3
aiofiles==23.2.1
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.6.0
anthropic==0.3.11
anyio==3.7.1
appdirs==1.4.4
appnope @ file:///home/conda/feedstock_root/build_artifacts/appnope_1649077682618/work
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
async-timeout==4.0.2
attrs==23.1.0
autollm==0.1.8
backoff==2.2.1
beautifulsoup4==4.11.1
bleach==6.1.0
blinker==1.6.2
boto3==1.24.53
botocore==1.27.96
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.6
cohere==4.1.3
colorama==0.4.6
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1704278392174/work
contourpy==1.2.0
crewai==0.1.17
cryptography==41.0.7
curl-cffi==0.5.10
cycler==0.12.1
dataclasses-json==0.5.14
datasets==2.15.0
debugpy @ file:///Users/runner/miniforge3/conda-bld/debugpy_1695534356146/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
distlib==0.3.8
distro==1.9.0
dnspython==2.4.2
docx==0.2.4
docx2txt==0.8
duckduckgo_search==4.1.1
EbookLib==0.18
environs==9.5.0
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1700579780973/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1698579936712/work
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.19.1
ffmpy==0.3.1
filelock==3.12.2
fonttools==4.47.0
frozendict==2.4.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gradio==4.14.0
gradio_client==0.8.0
greenlet==2.0.2
grpcio==1.56.0
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.19.4
identify==2.5.33
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.1
iniconfig==2.0.0
install==1.3.5
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1703631823913/work
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1703263236586/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1696326070614/work
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1699283905679/work
jupyter_core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1704321426280/work
jupyterlab_pygments==0.3.0
kaleido==0.2.1
kiwisolver==1.4.5
lancedb==0.3.4
langchain==0.0.251
langchain-community==0.0.9
langchain-core==0.1.7
langchain-openai==0.0.2
langsmith==0.0.18
litellm==1.16.21
llama-index==0.9.27
llmware==0.1.14
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1660814786464/work
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
multitasking==0.0.11
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1697083700168/work
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.2
openai==0.27.7
openapi-schema-pydantic==1.2.4
orjson==3.9.10
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pdf2image==1.16.0
pdfminer.six==20221105
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1667297516076/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.5.0
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1701708255999/work
playwright==1.40.0
pluggy==1.3.0
pre-commit==3.4.0
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1702399386289/work
protobuf==4.23.4
psutil @ file:///Users/runner/miniforge3/conda-bld/psutil_1702833169699/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1642875951954/work
py==1.11.0
pyarrow==12.0.1
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==1.10.12
pydantic_core==2.14.6
pydeck==0.8.0
pydub==0.25.1
pyee==11.0.1
Pygments==2.15.1
pylance==0.8.17
pymilvus==2.3.0
pymongo==4.5.0
Pympler==1.0.1
pyparsing==3.1.1
PyPDF2==3.0.1
pytesseract==0.3.10
pytest==7.4.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
python-docx==0.8.11
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0.1
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1701783262640/work
ratelimiter==1.2.0.post0
referencing==0.30.1
regex==2023.6.3
requests==2.31.0
retry==0.9.2
rich==13.5.2
rpds-py==0.9.2
s3transfer==0.6.2
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
seaborn==0.13.1
semantic-version==2.10.0
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.19
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
starlette==0.27.0
streamlit==1.25.0
streamlit-chat==0.1.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.5.2
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1704212992681/work
transformers==4.35.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.25.0
validators==0.20.0
virtualenv==20.25.0
watchdog==3.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1700607916581/work
webencodings==0.5.1
websockets==11.0.3
Werkzeug==3.0.1
Wikipedia-API==0.6.0
word2number==1.1
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.2
yfinance==0.2.28
zipp==3.16.2
